Viz2Speech/app/TTS/requirements.txt
chatterbox-tts==0.1.1
transformers>=4.44.2,<=4.46.3
peft==0.11.1
setuptools==75.2.0
pandas
numpy
matplotlib
seaborn
scikit-learn
tqdm
ml_dtypes
datasets
bitsandbytes
safetensors
accelerate
fastapi
uvicorn
python-multipart
soundfile